Добавил dto, мапперы, di модули network и domain, remote datasource

+++ app/src/main/java/com/test/a2021_q4_tyukavkin/domain/entity/LoanConditions.kt
@@ -1,7 +1,9 @@
 package com.test.a2021_q4_tyukavkin.domain.entity
 
+import java.math.BigDecimal
+
 data class LoanConditions (
-    private val maxAmount: Float, //TODO Уточнить тип number
+    private val maxAmount: BigDecimal, //TODO Уточнить тип number
     private val percent: Double,
     private val period: Int
 )
